+++ cli/src/templates/default/hardhat/ts/mocha-ethers/ignition/modules/Counter.ts
@@ -1,12 +1,9 @@
 import { buildModule } from "@nomicfoundation/hardhat-ignition/modules";
 
-import { DiamondInspectFacet } from "@perfect-abstractions/compose/diamond/DiamondInspectFacet.sol";
-import { DiamondUpgradeFacet } from "@perfect-abstractions/compose/diamond/DiamondUpgradeFacet.sol";
-
 export default buildModule("CounterDiamondModule", (m) => {
   const counterFacet = m.contract("CounterFacet");
-  const inspectFacet = m.contract(DiamondInspectFacet);
-  const upgradeFacet = m.contract(DiamondUpgradeFacet);
+  const inspectFacet = m.contract("DiamondInspectFacet");
+  const upgradeFacet = m.contract("DiamondUpgradeFacet");
 
   const owner = m.getAccount(0);
   const diamond = m.contract("Diamond", [[counterFacet, inspectFacet, upgradeFacet], owner]);