+++ cli/CHANGELOG.md
@@ -1,6 +1,31 @@
 # Changelog
 
-## 0.1.0
+## 0.0.5
+
+### Patch Changes
+
+- 404aafc: fix deployment setup on hardhat
+
+## 0.0.4
+
+### Patch Changes
+
+- 84abb2d: add installation commands to docs & readme
+
+## 0.0.3
+
+### Patch Changes
+
+- 33519f4: fix: compiler versions & imports
+  docs: add local node command
+
+## 0.0.2
+
+### Patch Changes
+
+- a9f32fb: first publised release: core facet library (@perfect-abstraction/compose) and CLI (@perfect-abstraction/compose-cli)
+
+## 0.0.1
 
 - Initial Compose CLI foundation.
 - Added config-driven template registry with Foundry/Hardhat variants.